pyproject.toml
[project]
name = "chance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
]
